+++ include/mechanism_configuration/mechanism_configuration.hpp
@@ -4,10 +4,12 @@
 
 #pragma once
 
-#include <mechanism_configuration/aerosol_types.hpp>
 #include <mechanism_configuration/errors.hpp>
 #include <mechanism_configuration/mechanism.hpp>
 #include <mechanism_configuration/parse.hpp>
-#include <mechanism_configuration/types.hpp>
+#include <mechanism_configuration/types/aerosol.hpp>
+#include <mechanism_configuration/types/emissions.hpp>
+#include <mechanism_configuration/types/reactions.hpp>
+#include <mechanism_configuration/types/species.hpp>
 #include <mechanism_configuration/validate.hpp>
 #include <mechanism_configuration/version.hpp>
