+++ include/mechanism_configuration/types/reactions.hpp
@@ -4,51 +4,12 @@
 
 #pragma once
 
-#include <mechanism_configuration/errors.hpp>
-
-#include <array>
-#include <optional>
 #include <string>
 #include <unordered_map>
 #include <vector>
 
 namespace mechanism_configuration::types
 {
-
-  struct Species
-  {
-    std::string name;
-    std::optional<double> absolute_tolerance;
-    std::optional<double> diffusion_coefficient;
-    std::optional<double> molecular_weight;
-    std::optional<double> henrys_law_constant_298;
-    std::optional<double> henrys_law_constant_exponential_factor;
-    std::optional<double> n_star;
-    std::optional<double> density;
-    std::optional<std::string> tracer_type;
-    std::optional<double> constant_concentration;
-    std::optional<double> constant_mixing_ratio;
-    std::optional<bool> is_third_body;
-    /// @brief Unknown properties, prefixed with two underscores (__)
-    std::unordered_map<std::string, std::string> unknown_properties;
-  };
-
-  struct PhaseSpecies
-  {
-    std::string name;
-    std::optional<double> diffusion_coefficient;
-    std::optional<double> density;
-    /// @brief Unknown properties, prefixed with two underscores (__)
-    std::unordered_map<std::string, std::string> unknown_properties;
-  };
-
-  struct Phase
-  {
-    std::string name;
-    std::vector<PhaseSpecies> species;
-    /// @brief Unknown properties, prefixed with two underscores (__)
-    std::unordered_map<std::string, std::string> unknown_properties;
-  };
 
   struct ReactionComponent
   {
@@ -325,90 +286,6 @@ namespace mechanism_configuration::types
     std::vector<Tunneling> tunneling;
     std::vector<UserDefined> user_defined;
     std::vector<LambdaRateConstant> lambda_rate_constant;
-  };
-
-  // ── Emissions configuration types ─────────────────────────────────────────
-
-  struct SpeciesMapping
-  {
-    std::string inventory_species;
-    std::string mechanism_species;
-    double scaling_factor{ 1.0 };
-  };
-
-  struct SpeciesMap
-  {
-    std::string name;
-    std::vector<SpeciesMapping> mappings;
-  };
-
-  struct Inventory
-  {
-    std::string name;
-    std::string directory;
-    std::string file_pattern;
-    std::string convention;
-  };
-
-  enum class SourceMode
-  {
-    Offline,
-  };
-
-  enum class SourceType
-  {
-    Anthropogenic,
-    Fire,
-    Biogenic,
-    Dust,
-    SeaSalt,
-    Lightning,
-  };
-
-  enum class TemporalInterpolation
-  {
-    Linear,
-    Nearest,
-    None,
-  };
-
-  enum class VerticalInjection
-  {
-    Surface,
-  };
-
-  struct SourceDescriptor
-  {
-    std::string name;
-    SourceMode mode{ SourceMode::Offline };
-    SourceType type{ SourceType::Anthropogenic };
-    std::string inventory;
-    std::string species_map;
-    TemporalInterpolation temporal_interpolation{ TemporalInterpolation::Linear };
-    VerticalInjection vertical_injection{ VerticalInjection::Surface };
-    int category{ 0 };
-    int hierarchy{ 1 };
-    double scaling_factor{ 1.0 };
-    std::string sector;
-    std::unordered_map<std::string, std::string> unknown_properties;
-  };
-
-  enum class RegriddingType
-  {
-    None,
-  };
-
-  struct Regridding
-  {
-    RegriddingType type{ RegriddingType::None };
-  };
-
-  struct EmissionsConfig
-  {
-    std::vector<Inventory> inventories;
-    std::vector<SpeciesMap> species_maps;
-    Regridding regridding;
-    std::vector<SourceDescriptor> sources;
   };
 
 }  // namespace mechanism_configuration::types
