+++ src/detail/v1/aerosol_type_parsers.hpp
@@ -4,8 +4,9 @@
 
 #pragma once
 
-#include <mechanism_configuration/aerosol_types.hpp>
-#include <mechanism_configuration/types.hpp>
+#include <mechanism_configuration/types/aerosol.hpp>
+#include <mechanism_configuration/types/reactions.hpp>
+#include <mechanism_configuration/types/species.hpp>
 
 #include <yaml-cpp/yaml.h>
