+++ src/detail/v1/type_schema.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include <mechanism_configuration/errors.hpp>
-#include <mechanism_configuration/types.hpp>
+#include <mechanism_configuration/types/species.hpp>
 
 #include <yaml-cpp/yaml.h>
